requirements.txt
hikari[speedups]
hikari-lightbulb
uvloop
python-dotenv
asyncpg